cloud/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-socketio
# TODO: Add any other necessary dependencies 